自己位置推定&ジャイロによる目的地追従を追加中

+++ sdk/workspace/BeachFlags/detection/detectCount.h
@@ -11,7 +11,7 @@ public:
 		bool detect() override;
 
 private:
-    int mTemp;
+    double mTemp;
 	int mCount;
 
 };
